Trait dex_swap_router::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type StablePoolId: Parameter + Codec + Copy + Ord + AtLeast32BitUnsigned + Zero + One + Default;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Codec + Default + Copy + MaybeSerializeDeserialize + Debug + MaxEncodedLen + From<AssetBalance> + Into<AssetBalance> + TypeInfo;
    type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen;
    type GeneralAmm: ExportDexGeneral<<Self as Config>::AccountId, Self::CurrencyId>;
    type StableAmm: StableAmmApi<Self::StablePoolId, Self::CurrencyId, <Self as Config>::AccountId, Self::Balance>;
    type GeneralWeightInfo: DexGeneralWeightInfo;
    type StableWeightInfo: DexStableWeightInfo;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type StablePoolId: Parameter + Codec + Copy + Ord + AtLeast32BitUnsigned + Zero + One + Default

type Balance: Parameter + Member + AtLeast32BitUnsigned + Codec + Default + Copy + MaybeSerializeDeserialize + Debug + MaxEncodedLen + From<AssetBalance> + Into<AssetBalance> + TypeInfo

type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen

type GeneralAmm: ExportDexGeneral<<Self as Config>::AccountId, Self::CurrencyId>

type StableAmm: StableAmmApi<Self::StablePoolId, Self::CurrencyId, <Self as Config>::AccountId, Self::Balance>

type GeneralWeightInfo: DexGeneralWeightInfo

Weight information for extrinsics in this pallet.

type StableWeightInfo: DexStableWeightInfo

type WeightInfo: WeightInfo